Add a tarball handler
